Created API, improved services, added more custom exceptions

+++ src/main/java/pl/edu/pwr/concertbooker/service/row/RowService.java
@@ -2,13 +2,12 @@ package pl.edu.pwr.concertbooker.service.row;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import pl.edu.pwr.concertbooker.exception.custom.CannotEditVenueWithExistingEventsException;
 import pl.edu.pwr.concertbooker.exception.custom.EntityNotFoundException;
 import pl.edu.pwr.concertbooker.model.Row;
 import pl.edu.pwr.concertbooker.model.Sector;
 import pl.edu.pwr.concertbooker.repository.RowRepository;
-import pl.edu.pwr.concertbooker.service.interfaces.IRowService;
-import pl.edu.pwr.concertbooker.service.interfaces.ISectorService;
-import pl.edu.pwr.concertbooker.service.interfaces.IVenueService;
+import pl.edu.pwr.concertbooker.service.interfaces.*;
 import pl.edu.pwr.concertbooker.service.row.dto.CreateRowDto;
 import pl.edu.pwr.concertbooker.service.row.dto.RowInfoDto;
 import pl.edu.pwr.concertbooker.service.row.dto.UpdateRowDto;
@@ -21,18 +20,30 @@ import java.util.Optional;
 public class RowService implements IRowService {
     RowRepository rowRepository;
     ISectorService sectorService;
+    IVenueService venueService;
+    IVenueUsageService venueUsageService;
     @Override
-    public void addRow(CreateRowDto rowDto) throws EntityNotFoundException {
+    public Row addRow(CreateRowDto rowDto) throws EntityNotFoundException, CannotEditVenueWithExistingEventsException {
         var sector = sectorService.getSector(rowDto.getSectorId());
+        var venue = venueService.getVenue(sector.getVenue().getId());
+        if (!venueUsageService.getEventsForVenueId(venue.getId()).isEmpty()) {
+            throw new CannotEditVenueWithExistingEventsException();
+        }
         Row row = new Row();
         row.setName(rowDto.getName());
         row.setSector(sector);
 
         rowRepository.save(row);
+        return row;
     }
 
     @Override
-    public void updateRow(UpdateRowDto rowDto) throws EntityNotFoundException {
+    public void updateRow(UpdateRowDto rowDto) throws EntityNotFoundException, CannotEditVenueWithExistingEventsException {
+        var sector = sectorService.getSector(rowDto.getSectorId());
+        var venue = venueService.getVenue(sector.getVenue().getId());
+        if (!venueUsageService.getEventsForVenueId(venue.getId()).isEmpty()) {
+            throw new CannotEditVenueWithExistingEventsException();
+        }
         Optional<Row> rowOptional = rowRepository.findById(rowDto.getId());
         if (rowOptional.isPresent()){
             Row row = rowOptional.get();
@@ -60,9 +71,14 @@ public class RowService implements IRowService {
     }
 
     @Override
-    public void deleteRowById(long id) throws EntityNotFoundException {
+    public void deleteRowById(long id) throws EntityNotFoundException, CannotEditVenueWithExistingEventsException {
         Optional<Row> rowOptional = rowRepository.findById(id);
         if (rowOptional.isPresent()) {
+            var sector = sectorService.getSector(rowOptional.get().getSector().getId());
+            var venue = venueService.getVenue(sector.getVenue().getId());
+            if (!venueUsageService.getEventsForVenueId(venue.getId()).isEmpty()) {
+                throw new CannotEditVenueWithExistingEventsException();
+            }
             rowRepository.deleteById(id);
         } else {
             throw new EntityNotFoundException(id);
